There is an odd conflict in a jenkins file I don't remember touching.

+++ groups/bsl/bslstl/bslstl_list.t.cpp
@@ -4453,7 +4453,7 @@ void TestDriver<TYPE,ALLOC>::testInsert()
             char *v = 0;
 
             x.insert(X.begin(), n, 0);  // Literal null, acts like an int.
-            ASSERT((int) X.size()  == n);
+            ASSERT(X.size()  == (size_t)n);
             ASSERT(X.front() == v);
             ASSERT(X.back()  == v);
         }
@@ -6236,7 +6236,7 @@ void TestDriver<TYPE,ALLOC>::testConstructor()
             list<char*, ALLOC> x(n, 0);  // Literal null, acts like an int.
             list<char*, ALLOC>& X = x;
 
-            ASSERT((int) X.size()  == n);
+            ASSERT(X.size()  == (size_t)n);
             ASSERT(X.front() == v);
             ASSERT(X.back()  == v);
         }
@@ -8715,6 +8715,8 @@ int main(int argc, char *argv[])
         if (verbose) printf("\nusageExample2\n");
         usageExample2(veryVerbose);
 
+        // Erase output files.
+
         remove("star_data1.txt");
         remove("star_data2.txt");
 
@@ -9894,7 +9896,7 @@ int main(int argc, char *argv[])
 }
 
 // ----------------------------------------------------------------------------
-// Copyright (C) 2012 Bloomberg L.P.
+// Copyright (C) 2013 Bloomberg L.P.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to
